fix: Also update custom_name if it matches name
Further digging in db.py shows that `name` is actually a computed property
that is backed by the `_computed_name` field in the DB. The logic to
compute `custom_name` and `gateway_slug` automatically populates these
fields if not set in the initial payload, so on subsequent updates, these
values are set and act as an override to the value being set when `name`
gets updated.

The logic in this fix is potentially problematic if a user did explicitly
set the `custom_name` field to match `name` and for some reason wants to
update `name` but NOT `custom_name`. This would not actually make sense to
do, though, so I think this is fairly correct at the moment.

https://github.com/IBM/mcp-context-forge/issues/1485
Branch: ToolNameUpdate-1485

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>