feat: Lazy initialization (#1492)
* refactor: Remove proactive imports from top-level __init__.py

Branch: LazyImports

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* feat: Add client_mode to settings

This allows client-side apps such as the CLI to use the settings without
all of the server-side initialization and validation.

Branch: LazySettings

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* feat: Create a lazy wrapper around the `settings` singleton

The underlying object will only be instantiated when an attribute is
accessed. This will enable non-server uses of the library to carefully set
up settings before triggering the creation of the singleton.

NOTE: We still need to plumb-out all of the import-time settings use!

Branch: LazySettings

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* fix: Fix warning skipping for settings in client mode

Branch: LazyImports

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* test: Re-clear cache after testing LRU with the Settings class mocked

Branch: LazyImports

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* fix: Fix flake8 docstring issues

Also remove the ability to give an arbitrary get_settings function to the
lazy wrapper. This for some reason was not passing flake8 even with the
docstring fixed.

Branch: LazyImports

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* fix: Correct type annotation for get_settings kwargs parameter

The **kwargs type hint was incorrectly specified as Dict[str, Any],
which implies each value is a Dict. The correct annotation is Any,
indicating each value in kwargs can be of any type.

Signed-off-by: Mihai Criveti <crivetimihai@gmail.com>

---------

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>
Signed-off-by: Mihai Criveti <crivetimihai@gmail.com>
Co-authored-by: Mihai Criveti <crivetimihai@gmail.com>